% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube.R
\name{datacube}
\alias{datacube}
\title{Create Datacube Class}
\usage{
datacube(id)
}
\arguments{
\item{id}{character containing the collection name to be passed to
openeo::processes()$load_collection() (https://processes.openeo.org/#load_collection).
This is understood as the **starting point** when working with openeowrap.}
}
\value{
datacube
}
\description{
Defines the class datacube based on collection name
}
