% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter.datacube}
\title{Filter Datacube}
\usage{
filter(
  .data = NULL,
  .bands = NULL,
  .wavelength = NULL,
  .condition = NULL,
  .dimension = NULL,
  .context = NULL
)

\method{filter}{datacube}(
  .data = NULL,
  .bands = NULL,
  .wavelength = NULL,
  .condition = NULL,
  .dimension = NULL,
  .context = NULL
)
}
\arguments{
\item{.data}{datacube object from openeowrap.}

\item{.bands}{For **filter_bands** : character. A list of band names. Either
the unique band name (metadata field name in bands) or one of the common band
names (metadata field common_name in bands). If the unique band name and the
common name conflict, the unique band name has a higher priority.
The order of the specified array defines the order of the bands in the data
cube. If multiple bands match a common name, all matched bands are included in
the original order.}

\item{.wavelength}{For **filter_bands** : character. A list of sub-lists with
each sub-list consisting of two elements. The first element is the minimum
wavelength and the second element is the maximum wavelength. Wavelengths are
specified in micrometers (μm). The order of the specified array defines the
order of the bands in the data cube. If multiple bands match the wavelengths,
all matched bands are included in the original order.}

\item{.condition}{For **filter_labels** : logical. A condition that is evaluated
against each dimension label in the specified dimension. A dimension label and
the corresponding data is preserved for the given dimension, if the condition
returns true.}

\item{.dimension}{For **filter_labels** (optional) : character The name of the dimension
to filter on. Fails with a DimensionNotAvailable exception if the specified
dimension does not exist.}

\item{.context}{For **filter_labels** (optional) : any Additional data to be passed to the condition.}
}
\value{
datacube
}
\description{
Filter datacube wraps the filter_bands and filter_labels function
into a simulated dplyr's \code{\link[dplyr]{filter}}.
}
\details{
if args bands and/or wavelengths are defined,
openeo::list_processes()$filter_bands
will be invoked. If condition and dimensions are defined,
openeo::list_processes()$filter_labels
will be called.
}
\examples{
library(openeowrap)
dc = datacube(id = "SENTINEL_5P_L2")

dc_no2 <- dc \%>\% filter(.bands = "NO2")
}
\seealso{
[openeo::list_processes()]
}
