% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.datacube}
\title{Rename Datacube Dimension}
\usage{
\method{rename}{datacube}(
  .data = NULL,
  ...,
  .source,
  .target,
  .p = openeo::processes(.con),
  .con = openeo::connect("openeo.cloud")
)
}
\arguments{
\item{.data}{datacube object from tidyopeneo.}

\item{...}{any parameter inherited from dplyr}

\item{.source}{The current name of the dimension. Fails with a
`DimensionNotAvailable` exception if the specified dimension does not exist.}

\item{.target}{A new Name for the dimension. Fails with a `DimensionExists`
exception if a dimension with the specified name exists.}

\item{.p}{processes available at .con}

\item{.con}{openeo connection}
}
\value{
datacube
}
\description{
Rename datacube dimension  wraps the rename_dimension(https://processes.openeo.org/#rename_dimension),
 function into a simulated dplyr's \code{\link[dplyr]{rename}}.
}
\examples{
library(tidyopeneo)

# bounding box
w = 6.09
s = 46.15
e = 6.99
n = 46.5

## time extent
date1 = "2018-07-01"
date2 = "2018-10-31"

dc = datacube("SENTINEL_5P_L2") \%>\%
   slice(.extent = list(west = w, south = s, east = e, north = n)) \%>\%
   slice(.extent = c(date1, date2)) \%>\%
   select(.bands = "NO2") \%>\%
   rename(.source = "spatial", .target = "space")
}
\seealso{
[openeo::list_processes()]
}
