% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select Datacube}
\arguments{
\item{.data}{datacube object from tidyopeneo.}

\item{.bands}{character. A list of band names. Either
the unique band name (metadata field name in bands) or one of the common band
names (metadata field common_name in bands). If the unique band name and the
common name conflict, the unique band name has a higher priority.
The order of the specified array defines the order of the bands in the data
cube. If multiple bands match a common name, all matched bands are included in
the original order.}

\item{.wavelength}{character. A list of sub-lists with
each sub-list consisting of two elements. The first element is the minimum
wavelength and the second element is the maximum wavelength. Wavelengths are
specified in micrometers. The order of the specified array defines the
order of the bands in the data cube. If multiple bands match the wavelengths,
all matched bands are included in the original order.}
}
\value{
datacube
}
\description{
Select datacube wraps the filter_bands function
into a simulated dplyr's \code{\link[dplyr]{select}}.
}
\examples{
library(tidyopeneo)
dc = datacube(id = "SENTINEL_5P_L2")

dc_no2 <- dc \%>\% select(.bands = "NO2")
}
\seealso{
[openeo::list_processes()]
}
