% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{summarize.datacube}
\title{Summarize Datacube}
\usage{
\method{summarize}{datacube}(
  .data = NULL,
  ...,
  .reducer = NULL,
  .dimension = NULL,
  .context = NULL,
  .p = openeo::processes(.con),
  .con = NULL
)
}
\arguments{
\item{.data}{datacube object from tidyopeneo.}

\item{...}{any parameter inherited from dplyr}

\item{.reducer}{character or function. A reducer to apply on the specified dimension.
A reducer is a single process such as "mean or a set of processes, which
computes a single value for a list of values, see the category 'reducer'
for such processes.}

\item{.dimension}{The name of the dimension over which to reduce. Fails with
a `DimensionNotAvailable` exception if the specified dimension does not exist.}

\item{.context}{(optional) Additional data to be passed to the reducer (optional).}

\item{.p}{(optional) processes available at .con}

\item{.con}{(optional) openeo connection. Default to NULL}
}
\value{
datacube
}
\description{
Summarize datacube wraps the reduce_dimension(https://processes.openeo.org/#reduce_dimension),
 function into a simulated dplyr's \code{\link[dplyr]{summarize}}.
}
\examples{
library(tidyopeneo)
con = connect(host = "https://openeo.cloud")

# bounding box
w = 6.09
s = 46.15
e = 6.99
n = 46.5

## time extent
date1 = "2018-07-01"
date2 = "2018-10-31"

dc = datacube("SENTINEL_5P_L2") \%>\%
   filter(.extent = list(west = w, south = s, east = e, north = n)) \%>\%
   filter(.extent = c(date1, date2)) \%>\%
   select(.bands = "NO2") \%>\%
   summarize(.dimension = "t", .reducer = mean)
}
\seealso{
[openeo::list_processes()]
}
