% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube.R
\name{datacube}
\alias{datacube}
\title{Create Datacube Class}
\usage{
datacube(id, data = NULL, .p = openeo::processes(.con), .con = NULL)
}
\arguments{
\item{id}{character containing the collection name to be passed to
openeo::processes()$load_collection() (https://processes.openeo.org/#load_collection).
This is understood as the **starting point** when working with tidyopeneo. (optional)}

\item{data}{ProcessNode datacube from openeo (optional)}

\item{.p}{(optional) processes available at .con}

\item{.con}{(optional) character link to openeo connection. Default to NULL}
}
\value{
datacube
}
\description{
Defines the class datacube based on collection name
}
\details{
either data or id must be called, and never both of them. Using id will call an
openEO connection and create a datacube object (tidyopeneo object) with the given collection.
Defining data is recommended to use once you need to apply a processes that is not included in
tidyopeneo and you want to come back to use tidyopeneo functions.
}
