% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter.datacube}
\title{Filter Datacube}
\usage{
\method{filter}{datacube}(
  .data = NULL,
  ...,
  .condition = NULL,
  .dimension = NULL,
  .context = NULL,
  .p = openeo::processes(.con),
  .con = openeo::connect("openeo.cloud")
)
}
\arguments{
\item{.data}{datacube object from tidyopeneo.}

\item{...}{any parameter inherited from dplyr}

\item{.condition}{logical. A condition that is evaluated
against each dimension label in the specified dimension. A dimension label and
the corresponding data is preserved for the given dimension, if the condition
returns true.}

\item{.dimension}{(optional) : character The name of the dimension
to filter on. Fails with a DimensionNotAvailable exception if the specified
dimension does not exist.}

\item{.context}{(optional) : any Additional data to be passed to the condition.}

\item{.p}{processes available at .con}

\item{.con}{openeo connection}
}
\value{
datacube
}
\description{
Filter datacube wraps the filter_labels function
into a simulated dplyr's \code{\link[dplyr]{filter}}.
}
\examples{
# TODO
}
\seealso{
[openeo::list_processes()]
}
