% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\alias{slice.datacube}
\title{Slice Datacube}
\usage{
\method{slice}{datacube}(
  .data = NULL,
  ...,
  n = 1,
  prop = NULL,
  .p = openeo::processes(.con),
  .con = NULL
)
}
\arguments{
\item{.data}{datacube object from tidyopeneo.}

\item{...}{any parameter inherited from dplyr}

\item{n}{integer. Provide either n, the number of days, or prop, the proportion of available days to select.
If neither are supplied, n = 1 will be used. If a negative value of n or prop is provided, the specified
number or proportion of days will be deselected.

If n is greater than the number of days available in the collection or datacube (or prop > 1), the result
will be truncated to the group size. If the proportion of a group size does not yield an integer number of
rows, the absolute value of prop * ndays(.data) is rounded down.}

\item{prop}{numeric Provide either n, the number of days, or prop, the proportion of available days to select.
If neither are supplied, n = 1 will be used. If a negative value of n or prop is provided, the specified
number or proportion of days will be deselected.

If n is greater than the number of days available in the collection or datacube (or prop > 1), the result
will be truncated to the group size. If the proportion of a group size does not yield an integer number of
rows, the absolute value of prop * ndays(.data) is rounded down.}

\item{.p}{(optional) processes available at .con}

\item{.con}{(optional) openeo connection. Default to NULL}
}
\value{
datacube
}
\description{
Slice datacube wraps the filter_temporal, just as filter.datacube does.
into a simulated dplyr's \code{\link[dplyr]{slice}}. The "n" from dplyr's slice is
interpreted as "days".
}
\examples{
library(tidyopeneo)
con = connect(host = "https://openeo.cloud")

dc = datacube("TERRASCOPE_S5P_L3_NO2_TD_V1")

dc1 = dc \%>\%
    filter(.extent = c("2021-01-01", "2021-03-03")) \%>\%
    slice(n = 3)

dc2 = dc \%>\% slice(n = -30)

dc3 = dc \%>\%
    filter(.extent = c("2021-05-01", "2021-09-31")) \%>\%
    slice(prop = 0.55)
}
\seealso{
[openeo::list_processes()]
}
